Created bash script to quickly download all revelant data.

+++ weather_api.py
@@ -1,6 +1,9 @@
 #!/usr/bin/python3
 
 import pandas as pd
+import subprocess
+import shlex
+from platform import system
 
 __all__ = ["get_data"]
 
@@ -27,14 +30,18 @@ def delete_mostly_nan_columns(df):
 
 def normalize_particular_columns(df, column_names):
     for col_name in column_names:
-        df.iloc[:,col_name]=(df.iloc[:,col_name]-df.iloc[:,col_name].mean())/df.iloc[:,col_name].std()
+        df[col_name]=(df[col_name]-df[col_name].mean())/df[col_name].std()
 
-def get_data():    
+def get_data(city_name):    
     columns_names = read_column_names('./data/header.txt')
-    df_smt = pd.read_csv('./data/s_m_t_2020.csv', names=columns_names[1:], index_col=False)
-
+    if system() is not "Linux":
+        sys.exit("ONLY LINUX")
+    subprocess.call(shlex.split(f'./get_files.sh {city_name}'))
+    df_smt = pd.read_csv(f'./data/s_m_t_{city_name}.csv', names=columns_names[1:], index_col=False)
     delete_mostly_nan_columns(df_smt)
-    columns_to_normalize = [[4],[5],[6],[7],[8],[9],[10],[11],[12]]
+    columns_to_normalize = ['Średnia dobowa temperatura [°C]', 'Status pomiaru TEMP',
+       'Status pomiaru WLGS', 'Średnia dobowa prędkość wiatru [m/s]',
+       'Status pomiaru FWS', 'Średnie dobowe zachmurzenie ogólne [oktanty]',
+       'Status pomiaru NOS']
     normalize_particular_columns(df_smt, columns_to_normalize)
-    print(df_smt.head(10))
     return df_smt
